docs: add screenshots of data insertion in videogames_dw and fix the queries

+++ Data Warehouse/src/analytic_queries.sql
@@ -1,5 +1,5 @@
 -- 1. Rata-rata rating video game per tahun
-SELECT y.year_id AS year, ROUND(AVG(f.rating), 2) AS average_rating
+SELECT y.year_id AS year, ROUND(AVG(f.rating)::numeric, 2) AS average_rating
 FROM fact_videogame_rating f
 JOIN dim_year y ON f.year_id = y.year_id
 GROUP BY y.year_id
@@ -14,7 +14,7 @@ GROUP BY d.director_name, y.year_id
 ORDER BY d.director_name, y.year_id;
 
 -- 3. Tren rating game yang disutradarai oleh director
-SELECT d.director_name, y.year_id AS year, ROUND(AVG(r.rating), 2) AS avg_rating
+SELECT d.director_name, y.year_id AS year, ROUND(AVG(r.rating)::numeric, 2) AS avg_rating
 FROM fact_director_credit f
 JOIN dim_director d ON f.director_id = d.director_id
 JOIN dim_year y ON f.year_id = y.year_id
@@ -23,7 +23,7 @@ GROUP BY d.director_name, y.year_id
 ORDER BY d.director_name, y.year_id;
 
 -- 4.  Sutradara dengan Rata-Rata Rating Tertinggi
-SELECT d.director_name, ROUND(AVG(r.rating), 2) AS avg_rating
+SELECT d.director_name, ROUND(AVG(r.rating)::numeric, 2) AS avg_rating
 FROM fact_director_credit f
 JOIN dim_director d ON f.director_id = d.director_id
 JOIN fact_videogame_rating r ON f.videogame_id = r.videogame_id AND f.year_id = r.year_id
